pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pustakalaya"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "textual>=0.70.0",
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.29.0",
    "jinja2>=3.1.0",
    "ebooklib>=0.18",
    "pymupdf>=1.24.0",
    "rarfile>=4.1",
    "Pillow>=10.3.0",
    "textual-image[textual]>=0.6.0",
    "watchdog>=4.0.0",
    "python-pam>=2.0.2",
    "black>=26.3.1",
]

[project.scripts]
pustakalaya = "pustakalaya.tui.app:main"
pustakalaya-web = "pustakalaya.web.app:main"

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "httpx>=0.27",
    "anyio[trio]>=4.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
